fastapi/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
azure-ai-formrecognizer==3.2.0
azure-common==1.1.28
azure-core==1.26.4
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cssselect==1.2.0
cycler==0.11.0
docopt==0.6.2
fastapi==0.95.1
fonttools==4.39.4
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-resources==5.12.0
isodate==0.6.1
joblib==1.2.0
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
msrest==0.7.1
multidict==6.0.4
num2words==0.5.12
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.7
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
plotly==5.14.1
pydantic==1.10.7
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
readability-lxml==0.8.1
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
zipp==3.15.0
